+++ R/utils.R
@@ -1,3 +1,0 @@
-iapply <- function(X, IND, FUN, ...) {
-  replace(X, IND, lapply(X[IND], FUN, ...))
-}